pass 1 (swap dynamic and physical inheritances, resolve compiletime errors, basic architecture down)
