+++ Polytoria/scripts/client/ui/menu/views/UIMenuOverview.cs
@@ -60,7 +60,7 @@ public sealed partial class UIMenuOverview : UIMenuViewBase
 	private void OnRespawn()
 	{
 		Menu.CoreUI.GameMenu.HideMenu();
-		Menu.CoreUI.Root.Players.LocalPlayer.Kill();
+		Menu.CoreUI.Root.Players.LocalPlayer.Character?.Kill();
 	}
 
 	private void OnScreenshot()
@@ -113,7 +113,7 @@ public sealed partial class UIMenuOverview : UIMenuViewBase
 			_userAvatarImage.LoadResource();
 		}
 
-		if (Menu.CoreUI.Service.CanRespawn)
+		if (Menu.CoreUI.Service.CanRespawn && Menu.CoreUI.Root.Players.LocalPlayer.Character != null)
 		{
 			_respawnButton.Modulate = new(1, 1, 1, 1f);
 			_respawnButton.MouseFilter = MouseFilterEnum.Stop;
