+++ Polytoria/scripts/datamodel/Player.cs
@@ -27,31 +27,15 @@ public sealed partial class Player : NPC
 	private const double MaxAFKTime = 60 * 15;
 	private const float CameraHeight = 2f;
 	public const string CreatorHeadScene = "res://scenes/creator/livecollab/head.tscn";
-	public const string BubbleChatScene = "res://scenes/client/spatial/chat/bubble_chat.tscn";
 	public const string BadgeImageDirPath = "res://assets/textures/client/ui/playerlist/badges/";
 	private static readonly Dictionary<string, string> _badgePathCache = [];
 	private bool _isReady = false;
-	internal bool ClimbDebounce = false;
-	internal bool JustFinishedClimbing = false;
 	internal bool IsMoving = false;
 	internal IPlayerMovement? PlayerMovement;
-	internal Vector3 LastVelocity;
-	internal Vector3 ExternalVelocity;
 
 	private float _respawnTime = 5.0f;
-	private bool _canMove = true;
-	private float _sprintSpeed;
-	private float _stamina = 0;
-	private float _maxStamina = 3;
-	private bool _useStamina = true;
-	private float _staminaRegen = 1.2f;
-	private float _staminaBurn = 1.2f;
-	private bool _keepInventory = false;
-	private bool _useHeadTurning = false;
 	private int _userID;
-	private bool _useBubbleChat = true;
 	private bool _autoLoadAppearance = true;
-	private bool _allowAnimationWhileMoving = false;
 	private PlayerMovementModeEnum _movementMode = PlayerMovementModeEnum.Default;
 	private PlayerRotationModeEnum _rotationMode = PlayerRotationModeEnum.Automatic;
 	private Team? _team;
@@ -65,16 +49,15 @@ public sealed partial class Player : NPC
 
 	internal bool teleporting = false;
 
-	private BubbleChat _bubbleChat = null!;
 	private RemoteTransform3D _remoteCamAttach = null!;
 	internal Dynamic CamAttach = null!;
 	private Physical? _mouseHoveringOn;
 
-	private Vector3 DefaultSpawnLocation = new(0, 5, 0);
+	public Vector3 DefaultSpawnLocation = new(0, 5, 0);
 	internal event Action<APIUserInfo>? UserInfoReady;
 
 #if CREATOR
-	private bool _spawnedAtCreatorPos = false;
+	internal bool _spawnedAtCreatorPos = false;
 #endif
 
 	// internal peer ID
@@ -116,83 +99,6 @@ public sealed partial class Player : NPC
 	}
 
 	[Editable, ScriptProperty]
-	public bool CanMove
-	{
-		get => _canMove;
-		set
-		{
-			_canMove = value;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty]
-	public float SprintSpeed
-	{
-		get => _sprintSpeed;
-		set
-		{
-			_sprintSpeed = value;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty, SyncVar(Unreliable = true, AllowAuthorWrite = true)]
-	public float Stamina
-	{
-		get => _stamina;
-		set
-		{
-			_stamina = value;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty]
-	public float MaxStamina
-	{
-		get => _maxStamina;
-		set
-		{
-			_maxStamina = value;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty, ScriptLegacyProperty("StaminaEnabled")]
-	public bool UseStamina
-	{
-		get => _useStamina;
-		set
-		{
-			_useStamina = value;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty]
-	public float StaminaRegen
-	{
-		get => _staminaRegen;
-		set
-		{
-			_staminaRegen = value;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty]
-	public float StaminaBurn
-	{
-		get => _staminaBurn;
-		set
-		{
-			_staminaBurn = value;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty]
 	public float RespawnTime
 	{
 		get => _respawnTime;
@@ -204,57 +110,12 @@ public sealed partial class Player : NPC
 	}
 
 	[Editable, ScriptProperty]
-	public bool KeepInventory
-	{
-		get => _keepInventory;
-		set
-		{
-			_keepInventory = value;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty]
-	public bool UseHeadTurning
-	{
-		get => _useHeadTurning;
-		set
-		{
-			_useHeadTurning = value;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty]
-	public bool UseBubbleChat
-	{
-		get => _useBubbleChat;
-		set
-		{
-			_useBubbleChat = value;
-			_bubbleChat?.Visible = _useBubbleChat;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty]
 	public bool AutoLoadAppearance
 	{
 		get => _autoLoadAppearance;
 		set
 		{
 			_autoLoadAppearance = value;
-			OnPropertyChanged();
-		}
-	}
-
-	[Editable, ScriptProperty]
-	public bool AllowAnimationWhileMoving
-	{
-		get => _allowAnimationWhileMoving;
-		set
-		{
-			_allowAnimationWhileMoving = value;
 			OnPropertyChanged();
 		}
 	}
@@ -385,12 +246,6 @@ public sealed partial class Player : NPC
 	[ScriptProperty]
 	public bool IsLocal { get; private set; }
 
-	[SyncVar(AllowAuthorWrite = true), ScriptProperty]
-	public bool IsClimbing { get; internal set; }
-
-	[SyncVar(AllowAuthorWrite = true), ScriptProperty]
-	public Truss? ClimbingTruss { get; internal set; }
-
 	[SyncVar(ServerOnly = true)]
 	public bool IsReady
 	{
@@ -412,11 +267,11 @@ public sealed partial class Player : NPC
 	[ScriptProperty, SyncVar]
 	public NetworkService.ClientPlatformEnum UserPlatform { get; internal set; }
 
-	[ScriptProperty]
-	public Inventory Inventory => FindChild<Inventory>("Inventory")!;
+	[Attributes.Obsolete("Use Character.Inventory instead"), ScriptProperty]
+	public Inventory? Inventory => Character?.Inventory;
 
-	[Attributes.Obsolete("Use Inventory instead"), ScriptProperty]
-	public Inventory Backpack => Inventory;
+	[Attributes.Obsolete("Use Character.Inventory instead"), ScriptProperty]
+	public Inventory? Backpack => Character?.Inventory;
 
 	// Emotes visible in emote wheel
 	public static readonly string[] EmoteWheelList =
@@ -429,39 +284,6 @@ public sealed partial class Player : NPC
 		"agree",
 		"disagree",
 	];
-
-	// List of all emotes
-	public static readonly string[] EmoteList =
-	[
-		"wave",
-		"dance",
-		"helicopter",
-		"sit",
-		"point",
-		"agree",
-		"disagree",
-		"scream",
-		"dance2",
-		"disappointed",
-	];
-
-	// Oneshot emotes
-	public static readonly string[] OneShotEmoteList =
-	[
-		"wave",
-		"point",
-		"disagree",
-		"agree",
-		"scream",
-		"disappointed",
-	];
-
-	public override void InitGDNode()
-	{
-		base.InitGDNode();
-		CollisionLayers = 2;
-		CollisionMask = 3;
-	}
 
 	public override void Init()
 	{
@@ -476,32 +298,14 @@ public sealed partial class Player : NPC
 			inventory.NetworkParent = this;
 		}
 
-		Died.Connect(OnPlayerDied);
 		Root.Players.PropertyChanged.Connect(OnPlayersPropertyChanged);
-
-		_bubbleChat = Globals.CreateInstanceFromScene<BubbleChat>(BubbleChatScene);
-		_bubbleChat.TargetPlayer = this;
-		_bubbleChat.Visible = _useBubbleChat;
-		GDNode.AddChild(_bubbleChat, @internal: Node.InternalMode.Back);
-		excludedBoundNodes.Add(_bubbleChat);
 	}
 
 	public override void PreDelete()
 	{
 		Root.Input.GodotInputEvent -= OnInput;
-		Died.Disconnect(OnPlayerDied);
 		PlayerMovement = null!;
 		base.PreDelete();
-	}
-
-	private void OnPlayerTouched(Physical obj)
-	{
-		obj.InvokeTouched(this);
-	}
-
-	private void OnPlayerTouchEnded(Physical obj)
-	{
-		obj.InvokeTouchEnded(this);
 	}
 
 	public override void Ready()
@@ -512,7 +316,7 @@ public sealed partial class Player : NPC
 
 	private void UpdatePlrReady()
 	{
-		SetCollisionDisabled(!_isReady);
+		Character?.SetCollisionDisabled(!_isReady);
 		GDNode3D?.Visible = _isReady;
 	}
 
@@ -547,11 +351,11 @@ public sealed partial class Player : NPC
 	{
 		if (Root.Players.PlayerCollisionEnabled)
 		{
-			SetCollisionMask(2, true);
+			Character?.SetCollisionMask(2, true);
 		}
 		else
 		{
-			SetCollisionMask(2, false);
+			Character?.SetCollisionMask(2, false);
 		}
 	}
 
@@ -565,13 +369,6 @@ public sealed partial class Player : NPC
 		if (!IsLocal)
 		{
 			UpdateTransformTick(delta);
-			if (Root.Network.IsServer && !IsSitting)
-			{
-				CharBody3D.Velocity = LastVelocity;
-				CharBody3D.MoveAndSlide();
-				LastVelocity = Vector3.Zero;
-				ApplyPushForce();
-			}
 		}
 
 		if (!IsLocal || !IsReady)
@@ -585,26 +382,6 @@ public sealed partial class Player : NPC
 		if (Root.Environment.CurrentCamera?.Mode != Camera.CameraModeEnum.Scripted)
 		{
 			Root.Environment.CurrentCamera?.CameraProcess(delta);
-		}
-	}
-
-	internal void AddStaminaTick(double delta)
-	{
-		if (!UseStamina) { return; }
-		Stamina += (float)(delta * StaminaRegen);
-		if (Stamina > MaxStamina)
-		{
-			Stamina = MaxStamina;
-		}
-	}
-
-	internal void RemoveStaminaTick(double delta)
-	{
-		if (!UseStamina) { return; }
-		Stamina -= (float)(delta * StaminaBurn);
-		if (Stamina < 0)
-		{
-			Stamina = 0;
 		}
 	}
 
@@ -627,25 +404,9 @@ public sealed partial class Player : NPC
 		}
 	}
 
-	internal void ApplyPushForce()
-	{
-		for (int i = 0; i < CharBody3D.GetSlideCollisionCount(); i++)
-		{
-			KinematicCollision3D collision = CharBody3D.GetSlideCollision(i);
-
-			if (GetNetObjFromProxy((Node)collision.GetCollider()) is Physical body)
-			{
-				// Push the rigidbody
-				body.ApplyForceFromPlayer(-collision.GetNormal());
-			}
-		}
-	}
-
 	public override void PhysicsProcess(double delta)
 	{
-		base.PhysicsProcess(delta);
-
-		if (Root.SessionType != World.SessionTypeEnum.Client || !IsLocal || !IsReady) { return; }
+		if (Root.SessionType != World.SessionTypeEnum.Client || !IsLocal || !IsReady || Character == null) { return; }
 
 		if (Character is PolytorianModel pt && pt.Ragdolling)
 		{
@@ -665,33 +426,7 @@ public sealed partial class Player : NPC
 			_mouseHoveringOn.MouseEnter.Invoke();
 		}
 
-		if (FootFwdRaycast.IsColliding())
-		{
-			Node collider = (Node)FootFwdRaycast.GetCollider();
-			if (collider != null && GetNetObjFromProxy(collider) is Truss truss)
-			{
-				if (!IsClimbing)
-				{
-					if (!ClimbDebounce)
-					{
-						ClimbingTruss = truss;
-						IsClimbing = true;
-						Character?.PlayClimb();
-					}
-
-				}
-			}
-			else
-			{
-				EndClimb();
-			}
-		}
-		else
-		{
-			EndClimb();
-		}
-
-		if (Anchored)
+		if (Character.Anchored)
 		{
 			// just in case it's anchored cuz ragdoll
 			if (Character is PolytorianModel pt2 && pt2.Ragdolling == false)
@@ -702,20 +437,14 @@ public sealed partial class Player : NPC
 			return;
 		}
 
+		if (Character.IsSitting) UpdateCamera(delta);
+
 		Camera? cam = Root.Environment.CurrentCamera;
 
 		// Apply camera modifier if enabled
-		if (UseHeadTurning && cam != null && cam.Mode == Camera.CameraModeEnum.Follow && cam.Target == CamAttach)
+		if (Character.UseHeadTurning && cam != null && cam.Mode == Camera.CameraModeEnum.Follow && cam.Target == CamAttach)
 		{
-			Character?.ApplyCameraModifier(cam);
-		}
-
-		if (IsSitting)
-		{
-			// Add stamina while sitting
-			AddStaminaTick(delta);
-			UpdateCamera(delta);
-			return;
+			Character.ApplyCameraModifier(cam);
 		}
 
 		if (PlayerMovement != null)
@@ -725,27 +454,18 @@ public sealed partial class Player : NPC
 		}
 		else
 		{
-			IsMoving = Velocity.Length() > 0.01f;
+			IsMoving = Character.Velocity.Length() > 0.01f;
 		}
 
 		// Stop animation on move
-		if (IsMoving && !AllowAnimationWhileMoving)
+		if (IsMoving && !Character.AllowAnimationWhileMoving)
 		{
-			Character?.Animator?.StopAnimation();
+			Character.Animator?.StopAnimation();
 		}
 
 		AfkTick(delta);
 
-		ApplyPushForce();
-	}
-
-	internal void EndClimb()
-	{
-		if (!IsClimbing) { return; }
-		IsClimbing = false;
-		JustFinishedClimbing = true;
-		ClimbingTruss = null;
-		Character?.SetAnimSpeed(1);
+		Character.ApplyPushForce();
 	}
 
 	private void SendPing()
@@ -787,23 +507,25 @@ public sealed partial class Player : NPC
 			if (Root.Environment.CurrentCamera?.Mode == Camera.CameraModeEnum.Free)
 			{
 				Root.Environment.CurrentCamera.Mode = Camera.CameraModeEnum.Follow;
-				CanMove = true;
+				Character?.CanMove = true;
 			}
 			else
 			{
 				Root.Environment.CurrentCamera?.Mode = Camera.CameraModeEnum.Free;
-				CanMove = false;
+				Character?.CanMove = false;
 			}
 		}
 
-		if (IsDead) { return; }
+		if (Character == null) return;
+
+		if (Character.IsDead) { return; }
 
 		if (@event.IsActionPressed("jump"))
 		{
 			// Ignore jump command if is custom
 			if (MovementMode == PlayerMovementModeEnum.Scripted) return;
-			if (!CanMove) return;
-			Jump();
+			if (!Character.CanMove) return;
+			Character.Jump();
 		}
 		else if (@event.IsActionPressed("toggle_sprint"))
 		{
@@ -811,17 +533,12 @@ public sealed partial class Player : NPC
 		}
 		else if (@event.IsActionPressed("drop_tool"))
 		{
-			DropTool();
+			Character?.DropTool();
 		}
 	}
 
-	private async void OnPlayerDied()
+	internal async void OnPlayerDied()
 	{
-		if (IsLocal)
-		{
-			UnequipTool();
-		}
-		Velocity = Vector3.Zero;
 		if (!Root.Network.IsServer) return; // Respawn on server only
 
 		// Respawn on client
@@ -853,11 +570,6 @@ public sealed partial class Player : NPC
 		cam.FirstPersonEntered.Connect(OnFirstPersonEntered);
 		cam.FirstPersonExited.Connect(OnFirstPersonExited);
 
-		// Listen to touch events
-		Touched.Connect(OnPlayerTouched);
-		TouchEnded.Connect(OnPlayerTouchEnded);
-		EnableCanTouch();
-
 		// Disable auto update, this will be updated manually
 		AutoUpdateNetTransform = false;
 
@@ -886,85 +598,23 @@ public sealed partial class Player : NPC
 		SetCamRemoteAttachEnabled(false);
 	}
 
-	internal void PlayEmote(string emoteName)
-	{
-		if (IsSitting || IsDead) return;
-		if (!EmoteList.Contains(emoteName)) return;
-		bool isOneShot = false;
-		if (OneShotEmoteList.Contains(emoteName))
-		{
-			isOneShot = true;
-		}
-
-		Character?.Animator?.StopAnimation();
-		Character?.Animator?.StopOneShotAnimation();
-
-		if (isOneShot)
-		{
-			Character?.Animator?.PlayOneShotAnimation("emote_" + emoteName);
-		}
-		else
-		{
-			Character?.Animator?.PlayAnimation("emote_" + emoteName);
-		}
-	}
-
 	internal void InvokeChatted(string msg)
 	{
 		Chatted.Invoke(msg);
-	}
-
-	public override void Jump()
-	{
-		base.Jump();
-		if (IsClimbing)
-		{
-			EndClimb();
-			ClimbDebounce = true;
-		}
 	}
 
 	private void OnFirstPersonEntered()
 	{
 		if (Character == null) return;
 		Character.GDNode3D.Visible = false;
-		_bubbleChat.Visible = false;
+		Character._bubbleChat.Visible = false;
 	}
 
 	private void OnFirstPersonExited()
 	{
 		if (Character == null) return;
 		Character.GDNode3D.Visible = true;
-		_bubbleChat.Visible = true;
-	}
-
-	public void WrapToSpawnPoint()
-	{
-		if (Root.Environment.SpawnPoints.Count > 0)
-		{
-			Entity spawnpoint = ArrayUtils.GetRandom(Root.Environment.SpawnPoints);
-			Position = spawnpoint.Position + new Vector3(0, spawnpoint.Size.Y + 2.0f, 0);
-			Rotation = new(0, spawnpoint.Rotation.Y, 0);
-		}
-		else
-		{
-			Position = DefaultSpawnLocation;
-			Rotation = new(0, 0, 0);
-		}
-
-		// Spawn at custom position
-#if CREATOR
-		if (Root.Entry != null && Root.Entry.DebugSpawnPos != null)
-		{
-			if (!_spawnedAtCreatorPos)
-			{
-				_spawnedAtCreatorPos = true;
-				Position = Root.Entry.DebugSpawnPos.Value;
-				Rotation = Vector3.Zero;
-			}
-		}
-#endif
-		SendNetTransformReliable();
+		Character._bubbleChat.Visible = true;
 	}
 
 	[ScriptMethod]
@@ -1000,29 +650,6 @@ public sealed partial class Player : NPC
 	}
 
 	[ScriptMethod]
-	public void UnequipTool()
-	{
-		if (HoldingTool == null) return;
-		Rpc(nameof(NetUnequipTool), HoldingTool.NetworkedObjectID);
-	}
-
-	[NetRpc(AuthorityMode.Authority, CallLocal = true, TransferMode = TransferMode.Reliable)]
-	private async void NetUnequipTool(string networkID)
-	{
-		NetworkedObject? netObj = await Root.WaitForNetObjectAsync(networkID);
-
-		if (netObj == null) { return; }
-
-		Tool tool = (Tool)netObj;
-
-		Character?.SetBlendValue(CharacterModel.CharacterModelBlendEnum.ToolHoldRight, 0);
-		tool.Parent = Inventory;
-		HoldingTool = null;
-		tool.InvokeUnequipped();
-		InternalDetachTool();
-	}
-
-	[ScriptMethod]
 	public new void Respawn()
 	{
 		InternalSpawn();
@@ -1034,33 +661,33 @@ public sealed partial class Player : NPC
 		CopyInventory();
 
 		// Apply playerdefaults
-		MaxHealth = Root.PlayerDefaults.MaxHealth;
-		WalkSpeed = Root.PlayerDefaults.WalkSpeed;
-		SprintSpeed = Root.PlayerDefaults.SprintSpeed;
-		UseStamina = Root.PlayerDefaults.UseStamina;
-		Stamina = Root.PlayerDefaults.Stamina;
-		MaxStamina = Root.PlayerDefaults.MaxStamina;
-		StaminaRegen = Root.PlayerDefaults.StaminaRegen;
-		StaminaBurn = Root.PlayerDefaults.StaminaBurn;
-		JumpPower = Root.PlayerDefaults.JumpPower;
+		if (Character != null)
+		{
+			Character.MaxHealth = Root.PlayerDefaults.MaxHealth;
+			Character.WalkSpeed = Root.PlayerDefaults.WalkSpeed;
+			Character.SprintSpeed = Root.PlayerDefaults.SprintSpeed;
+			Character.UseStamina = Root.PlayerDefaults.UseStamina;
+			Character.Stamina = Root.PlayerDefaults.Stamina;
+			Character.MaxStamina = Root.PlayerDefaults.MaxStamina;
+			Character.StaminaRegen = Root.PlayerDefaults.StaminaRegen;
+			Character.StaminaBurn = Root.PlayerDefaults.StaminaBurn;
+			Character.JumpPower = Root.PlayerDefaults.JumpPower;
+			Character.KeepInventory = Root.PlayerDefaults.KeepInventory;
+			Character.UseHeadTurning = Root.PlayerDefaults.UseHeadTurning;
+			Character.UseBubbleChat = Root.PlayerDefaults.UseBubbleChat;
+			if (Character is PolytorianModel ptmodel)
+			{
+				ptmodel.StopRagdoll();
+			}
+			Character.Velocity = Vector3.Zero;
+			Character.ResetAppearance();
+			Character.WrapToSpawnPoint();
+			Character.Health = Character.MaxHealth;
+			Character.Anchored = false;
+		}
 		RespawnTime = Root.PlayerDefaults.RespawnTime;
-		KeepInventory = Root.PlayerDefaults.KeepInventory;
-		UseHeadTurning = Root.PlayerDefaults.UseHeadTurning;
-		UseBubbleChat = Root.PlayerDefaults.UseBubbleChat;
 		AutoLoadAppearance = Root.PlayerDefaults.AutoLoadAppearance;
 		MovementMode = Root.PlayerDefaults.MovementMode;
-
-		if (Character is PolytorianModel ptmodel)
-		{
-			ptmodel.StopRagdoll();
-		}
-		Velocity = Vector3.Zero;
-
-		ResetAppearance();
-		WrapToSpawnPoint();
-
-		Health = MaxHealth;
-		Anchored = false;
 
 		Rpc(nameof(NetRespawned));
 	}
@@ -1069,10 +696,12 @@ public sealed partial class Player : NPC
 	{
 		// Only allow this operation in server
 		if (!Root.Network.IsServer) return;
+		if (Character == null) return;
+		if (Character.Inventory == null) return;
 
-		if (!KeepInventory)
+		if (Character.KeepInventory)
 		{
-			foreach (Instance item in Inventory.GetChildren())
+			foreach (Instance item in Character.Inventory.GetChildren())
 			{
 				item.Delete();
 			}
@@ -1085,7 +714,7 @@ public sealed partial class Player : NPC
 				NetworkedObject a = item.Clone();
 				if (a is Instance i)
 				{
-					i.Parent = Inventory;
+					i.Parent = Character.Inventory;
 				}
 			}
 		}
@@ -1096,26 +725,9 @@ public sealed partial class Player : NPC
 	{
 		Respawned?.Invoke();
 
-		OverrideCanCollide = false;
-		UpdateCollision();
-		IsDead = false;
-	}
-
-	[ScriptMethod]
-	public void ResetAppearance()
-	{
-		ClearAppearance();
-		if (AutoLoadAppearance)
-		{
-			if (Root.Entry != null && Root.Entry.IsSoloTest)
-			{
-				LoadAppearance(1144);
-			}
-			else
-			{
-				LoadAppearance(UserID);
-			}
-		}
+		Character?.OverrideCanCollide = false;
+		Character?.UpdateCollision();
+		Character?.IsDead = false;
 	}
 
 	internal override bool TransformNetworkCheck(TransformPayloadDto newTransform)
