+++ Polytoria/scripts/datamodel/services/InsertService.cs
@@ -73,7 +73,7 @@ public sealed partial class InsertService : Instance
 		jumpSound.PlayInWorld = true;
 		jumpSound.SetNetworkAuthority(owner, false);
 
-		npc.JumpSound = jumpSound;
+		ptm.JumpSound = jumpSound;
 
 		jumpSound.LocalPosition = Vector3.Zero;
 		jumpSound.LocalRotation = Vector3.Zero;
